requirements.txt
llama-index
nest-asyncio
openai
pydantic 
pypdf
python-dotenv
streamlit
